Lots of changes
The handsOff and handsOn methods are now distinct Code objects, like in LSL6.

The arrow cursor is now a distinct object, as is required in SCI32.

If an argument is given in restart: the game now restarts immediately without the dialog, for things like death messages.

The pragmaFail: method and doVerb code have been simplified.

The icon bar initializes with Walk as the selected icon.

Ego is normalized at startRoom:, ensuring he's always normalized at the start of every room.

GameInit now initializes text font and color codes for messages.