Use Flags object to manage event flags
The game now uses the Flags system script for managing event flags. The good thing about this is that the gameFlags global does not need to be part of an array. When macro defines are supported, the Bset, Btst, and Bclr procedures will be replaced with equivalent defines.

Thanks to key ticks no longer being case-sensitive, this created a positive side-effect of them being usable in defines! As a result, the original Sierra defines for ENTER, TAB, and BACKSPACE in SYSTEM.SH are now used.

Hopefully, SCICompanion will add support for macro defines, so that all of Sierra's original defines can be used.