Can't use YesNoDialog in main script
It seemed like a good idea, but the YesNoDialog requires pointing it to a message resource. This doesn't work well when the resource is 0, since it'll assume you didn't specify one.

Also, the icon bar is now disabled at first, so that the game doesn't crash with "Not an Object" when the icon bar tries to claim an event.